Only require fetchable payload
When constructing some object with  with a payload T
extends APIBasePayload, only require a Partial<T> & V, where V contains
the keys needed to identify our object and fetch it via the API

Preserve get behavior and type. Keep track of when we fetch a particular
object, and when we get a key, if the value of they key is undefined and
when we've never fetched before, fetch. This may result in
false-positives if the underlying type may be undefined, but I think
that's never the case?